requirements-pathwise.txt
# Core AI/LLM dependencies
nvidia-nat[langchain]>=1.2.0
aiqtoolkit>=1.2.0
langchain>=0.1.0
langchain-nvidia-ai-endpoints
langchain-community

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# CORS Proxy dependencies
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0

# Document processing
PyMuPDF>=1.23.0
python-multipart>=0.0.6

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Vector store
faiss-cpu>=1.7.4

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: Performance
uvloop>=0.19.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0